// frontend/src/app/model/role-enum.ts
export enum RoleEnum{
  CAN_CREATE = "CAN_CREATE",
  CAN_UPDATE = "CAN_UPDATE",
  CAN_READ = "CAN_READ",
  CAN_DELETE = "CAN_DELETE",
  CAN_ADD_VACUUM = "CAN_ADD_VACUUM",
  CAN_REMOVE_VACUUM = "CAN_REMOVE_VACUUM",
  CAN_START_VACUUM = "CAN_START_VACUUM",
  CAN_STOP_VACUUM = "CAN_STOP_VACUUM",
  CAN_DISCHARGE_VACUUM = "CAN_DISCHARGE_VACUUM",
  CAN_SEARCH_VACUUM = "CAN_SEARCH_VACUUM"
}